+++ lib/src/models/rgb_color.dart
@@ -16,19 +16,13 @@ class RgbColor extends ColorModel {
   /// A color in the sRGB color space.
   ///
   /// [red], [green], and [blue] must all be `>= 0` and `<= 255`.
-  factory RgbColor(num red, num green, num blue) {
-    assert(red != null && red >= 0 && red <= 255);
-    assert(green != null && green >= 0 && green <= 255);
-    assert(blue != null && blue >= 0 && blue <= 255);
-
-    return RgbColor._internal(red, green, blue);
-  }
-
-  RgbColor._internal(
+  const RgbColor(
     this._red,
     this._green,
     this._blue,
-  );
+  ) : assert(_red != null && _red >= 0 && _red <= 255),
+      assert(_green != null && _green >= 0 && _green <= 255),
+      assert(_blue != null && _blue >= 0 && _blue <= 255);
 
   /// The red value of this color.
   ///